% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf_to_xml}
\alias{pdf_to_xml}
\title{Parse pdf document as XML.}
\usage{
pdf_to_xml(filename, first, last)
}
\arguments{
\item{filename}{pdf file to process}

\item{first}{first page to process}

\item{last}{last page to process}
}
\value{
an \code{xml_document} class object from package \code{xml2}
}
\description{
Parse pdf document as XML.
}
\examples{
unmd_pdf <- system.file(package = "trickypdf", "extdata", "pdf", "UN_Millenium_Declaration.pdf")
unmd_xml <- pdf_to_xml(filename = unmd_pdf)

cdu_manifesto_pdf <- system.file(package = "trickypdf", "extdata", "pdf", "cdu.pdf")
cdu_manifesto_xml <- pdf_to_xml(filename = cdu_manifesto_pdf)

unga_pdf <- system.file(package = "trickypdf", "extdata", "pdf", "N9586353.pdf")
unga_xml <- pdf_to_xml(filename = unga_pdf)
}
