% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_paragraphs}
\alias{restore_paragraphs}
\title{Restore paragraphs}
\usage{
restore_paragraphs(x, skipRegexCurrent = "^\\\\s*[•A-Z(]",
  skipRegexPrevious = "[\\\\.?!)]\\\\s*$")
}
\arguments{
\item{x}{a character vector}

\item{skipRegexCurrent}{a regex}

\item{skipRegexPrevious}{another regex}
}
\description{
Restore paragraphs in a character vector.
}
\details{
Reconstruct paragraphs from a character vector with line breaks and word-wraps.
The heuristic is as follows: If a line ends with a hyphenation and the next line
starts with a small letter, remove hyphen and concatenate word."
}
\examples{
vec <- c(
  "This is a sample text. We freq-",
  "quently encounter issues with bro-",
  "ken lines."
  )
restore_paragraphs(vec)
}
