% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickypdf.R
\docType{package}
\name{trickypdf}
\alias{trickypdf}
\alias{trickypdf-package}
\title{pdf2xml}
\description{
Turn pdf document into XML for further processing 
in a corpus preparation pipeline. The particular
focus of the package is to handle tilted lines, and 
to handle pages with two columens.
}
\references{
http://polmine.sowi.uni-due.de
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
