% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_paragraphs}
\alias{restore_paragraphs}
\title{Restore paragraphs}
\usage{
restore_paragraphs(x, skipRegexCurrent = "^\\\\s*[A-Z(]",
  skipRegexPrevious = "[\\\\.?!)]\\\\s*$")
}
\arguments{
\item{x}{a character vector}

\item{skipRegexCurrent}{a regex}

\item{skipRegexPrevious}{another regex}
}
\description{
Restore paragraphs
}
